refactor: create seperate Search component
