+++ src/components/Search/Search.styled.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const SearchRow = styled.form`
+    display: flex;
+`;
